change specifier and cast in sprintf()

+++ libs/compiler/inline_expression.h
@@ -141,7 +141,7 @@ static char *create_new_temp_identifier_name(size_t ident_table_size)
 		printf("calloc error\n");
 		return NULL;
 	}
-	sprintf(new_identifier_number_str, "%d", (int)ident_table_size); 
+	sprintf(new_identifier_number_str, "%zu", (size_t)ident_table_size); 
 
 	char *new_identifier_name = calloc(1, sizeof(char));
 	if (!new_identifier_name)
